+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to a modified version of [Semantic Versioning](./README
 ### Added
 - feat: Add setting to disable the `loginWithPassword` mutation.
 - feat: Add settings to set a WP authentication cookie on successful login.
+- feat: Add support for WPGraphQL for WooCommerce.
 
 ### Changed
 - dev: Trigger `wp_login` action on successful login.